refactored the whole project into multiple file for readability
