+++ .gitignore
@@ -1,0 +1,3 @@
+venv
+.idea
+__pycache__
